Implementing some of Lars scripts
I have implemented the script for correcting Lemmas and Features for words [den, det, de, dem, dom]. As it works right now, a specific field of a token line is updated to a specific value if both its form and upos corresponds to the predefined keys in the dictionary dictionary in the changes_by_upos_config.py file. It is possible to add any number of token forms/upos combinations to the dictionary. If we need to make similar changes (only dependent on upos and form) in the future, these changes can be added in the config dictionary).